Resolve amdgpu BO_USE_CURSOR flag regression
This change resolves a regression in cursor behavior on amdgpu hardware
introduced by https://crrev.com/c/6282179. Prior to the culprit CL, the
GRALLOC_USAGE_CURSOR flag was always ignored. After the culprit CL, the
flag was not ignored, which started causing failed buffer allocations
due to an unsupported combination of the BO_USE_CURSOR use flag with
the DRM_FORMAT_ABGR8888 buffer format. This caused the mouse cursor to
not be visible.

This change adds amdgpu-specific behavior to ignore the BO_USE_CURSOR
flag whenever it is combined with an unsupported format, which should
restore the original behavior and resolve the regression.

Bug: b/404428779
Change-Id: I223452d27bdbadbfb9b74d55166e4c7f6c967bb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/minigbm/+/6405359
Commit-Queue: Andrew Wolfers <aswolfers@chromium.org>
Reviewed-by: Dominik Behr <dbehr@chromium.org>
Tested-by: Andrew Wolfers <aswolfers@chromium.org>