mediatek: advertise protected buffer support on ChromeOS
There is a sandbox or SELinux problem that blocks opening the Mediatek
CMA file for some Chrome processes. The result is that protected content
playback fails to display. This is only seen in signed images. Test
images where we run tast, don't see the same issue.

This always advertises protected buffer support when the OS is not
Android, even if the DMA heap file cannot be opened.

BUG=b/439910775
TEST=Widevine L1 playback on Navi

Change-Id: Icbe93c416a3ee6ad5906e3d321c87e8b5f5c2139
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/minigbm/+/6939292
Reviewed-by: Dominik Behr <dbehr@chromium.org>
Tested-by: Nathan Hebert <nhebert@chromium.org>
Commit-Queue: Nathan Hebert <nhebert@chromium.org>