dma-heap: add a new internal backend
This backend allocates dma-bufs from dma-heaps. It is generic and only
supports blobs.

It is a full backend like other backends. But it is also internal in the
sense that it is declared in drv_helpers.h and is never auto-selected by
drv_create.

It is also the only driver that is not based on DRM/GEM. As a result,
drv->fd is dma-heap fd rather than a drm fd. bo->handle.u32 is a
dma-buf fd rather than a gem handle.

BUG=b:446040761
TEST=boot navi

Change-Id: If5e274b254652ed9cc5be6f3b9d0a3defc8bea44
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/minigbm/+/6967760
Reviewed-by: Yiwei Zhang <zzyiwei@google.com>
Auto-Submit: Chia-I Wu <olv@google.com>
Tested-by: Chia-I Wu <olv@google.com>
Commit-Queue: Chia-I Wu <olv@google.com>
Reviewed-by: Ryan Neph <ryanneph@google.com>